feat: remove third party flag and rename subjects -> privacy
