+++ pkg/report/output/output.go
@@ -9,9 +9,8 @@ import (
 	"github.com/bearer/curio/pkg/commands/process/settings"
 	"github.com/bearer/curio/pkg/flag"
 	"github.com/bearer/curio/pkg/report/output/dataflow"
-	"github.com/bearer/curio/pkg/report/output/subjects"
+	"github.com/bearer/curio/pkg/report/output/privacy"
 	"github.com/bearer/curio/pkg/report/output/summary"
-	"github.com/bearer/curio/pkg/report/output/third_party"
 	"github.com/google/uuid"
 
 	"github.com/bearer/curio/pkg/report/output/detectors"
@@ -71,8 +70,6 @@ func ReportCSV(report types.Report, output *zerolog.Event, config settings.Confi
 	switch config.Report.Report {
 	case flag.ReportPrivacy:
 		return getPrivacyReportCsvOutput(report, output, config)
-	case flag.ReportThirdParty:
-		return getThirdPartyReportCsvOutput(report, output, config)
 	}
 
 	return fmt.Errorf("csv not supported for report type: %s", config.Report.Report)
@@ -120,8 +117,6 @@ func getReportOutput(report types.Report, config settings.Config) (any, error) {
 		return getSummaryReportOutput(report, config)
 	case flag.ReportPrivacy:
 		return getPrivacyReportOutput(report, config)
-	case flag.ReportThirdParty:
-		return getThirdPartyReportOutput(report, config)
 	case flag.ReportStats:
 		return reportStats(report, config)
 	}
@@ -135,7 +130,7 @@ func getPrivacyReportCsvOutput(report types.Report, output *zerolog.Event, confi
 		return err
 	}
 
-	csvString, err := subjects.BuildCsvString(dataflow, config)
+	csvString, err := privacy.BuildCsvString(dataflow, config)
 	if err != nil {
 		return err
 	}
@@ -144,37 +139,13 @@ func getPrivacyReportCsvOutput(report types.Report, output *zerolog.Event, confi
 	return nil
 }
 
-func getPrivacyReportOutput(report types.Report, config settings.Config) ([]subjects.InventoryResult, error) {
+func getPrivacyReportOutput(report types.Report, config settings.Config) ([]privacy.InventoryResult, error) {
 	dataflow, err := getDataflow(report, config, true)
 	if err != nil {
 		return nil, err
 	}
 
-	return subjects.GetOutput(dataflow, config)
-}
-
-func getThirdPartyReportOutput(report types.Report, config settings.Config) ([]third_party.InventoryResult, error) {
-	dataflow, err := getDataflow(report, config, true)
-	if err != nil {
-		return nil, err
-	}
-
-	return third_party.GetOutput(dataflow, config)
-}
-
-func getThirdPartyReportCsvOutput(report types.Report, output *zerolog.Event, config settings.Config) error {
-	dataflow, err := getDataflow(report, config, true)
-	if err != nil {
-		return err
-	}
-
-	csvString, err := third_party.BuildCsvString(dataflow, config)
-	if err != nil {
-		return err
-	}
-
-	output.Msg(csvString.String())
-	return nil
+	return privacy.GetOutput(dataflow, config)
 }
 
 func getSummaryReportOutput(report types.Report, config settings.Config) (map[string][]summary.PolicyResult, error) {
