--- pkg/report/output/subjects/subjects_test.go
@@ -1,1 +1,0 @@
-package subjects
